Company Database Implementation
Created company table using a migration. Then used a factory and seeder to add 100 fake companies to the table. Then used the all laravel method to get all the companies and create a card for each of them on the landing page. Used echoes to display column data in the relevant places of the cards